Enhanced Error Handling and Scheduling Improvements
Error Handling:
Changed the condition to check if a result is None instead of using not result.
Updated the order of parameters in the SQL query for weather data.

Error Logging:
Added a try-except block in the reo() function to catch and log errors.
Errors are now sent to the admin via Telegram for better monitoring.

Scheduling:
Adjusted the schedule times for the check_schedule function to run at 08:00 and 11:30 daily.

Location Check:
Modified the condition to check if result_location_exist is None instead of checking its first element.
Added additional checks for the elements of result_location_exist to ensure they are not None.